backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# Google GenAI
google-genai==0.6.0

# Redis + async job queue
redis==5.0.4
arq==0.26.1

# Scraping
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2

# Stripe billing
stripe==9.9.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database (SQLite dev / PostgreSQL prod)
sqlalchemy==2.0.30
aiosqlite==0.20.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.1
tenacity==8.3.0
aiofiles==23.2.1

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0
